+++ KuGuanXiTong/ChuKu_addTaoJian_Dialog.cs
@@ -87,7 +87,7 @@ namespace KuGuanXiTong
                             DeliverySuiteRecords rwg = (DeliverySuiteRecords)o[0];
                             if (rwg.GoodsInfoId.Id.ToString().Equals(Convert.ToInt32(s.GoodsCode.Substring(0, 8)).ToString()))
                             {
-                                r.Cells[6].Value = Convert.ToInt32(r.Cells[6].Value) - Convert.ToInt32(s.Quantity);
+                                r.Cells[6].Value = Convert.ToInt32(r.Cells[6].Value) - Convert.ToDouble(s.Quantity);
                                 IList<StockOperationDetail> stockList = null;
                                 if (o[1] == null)
                                 {
