+++ KuGuanXiTong/ChuKu_addWeiXiu_Dialog.cs
@@ -84,7 +84,7 @@ namespace KuGuanXiTong
                             ServiceGoods rwg = (ServiceGoods)o[0];
                             if (rwg.GoodsBaseInfoId.Id.ToString().Equals(Convert.ToInt32(s.GoodsCode.Substring(0, 8)).ToString()))
                             {
-                                r.Cells[6].Value = Convert.ToInt32(r.Cells[6].Value) - Convert.ToInt32(s.Quantity);
+                                r.Cells[6].Value = Convert.ToInt32(r.Cells[6].Value) - Convert.ToDouble(s.Quantity);
                                 IList<StockOperationDetail> stockList = null;
                                 if (o[1] == null)
                                 {
